+++ hackerrank/practice/mathematics/number_theory/equations.cpp
@@ -15,7 +15,7 @@ int main_equations()
     for (uint64_t p : primes_list.GetPrimes())
     {
         if (p > n) break;
-        unsigned c = 0;
+        uint64_t c = 0;
         for (uint64_t t = n/p; t; t /= p)
             c += t;
         r *= TModular(2 * c + 1);
